+++ Changelog.md
@@ -2,6 +2,15 @@
 
 ## Gazebo 7.x.x (2017-xx-xx)
 
+1. Parallelize ODE physics with threaded islands parameter
+    * [Pull request 2775](https://bitbucket.org/osrf/gazebo/pull-requests/2775)
+
+1. Logical camera uses <topic>
+    * [Pull request 2777](https://bitbucket.org/osrf/gazebo/pull-requests/2777)
+
+1. Support off-diagonal inertia terms in bullet
+    * [Pull request 2757](https://bitbucket.org/osrf/gazebo/pull-requests/2757)
+
 1. Add option in gui.ini to disable the use of spacenav
     * [Pull request 2754](https://bitbucket.org/osrf/gazebo/pull-requests/2754)
 
@@ -18,6 +27,9 @@
 
 1. Fix orbiting view around heightmap
     * [Pull request 2688](https://bitbucket.org/osrf/gazebo/pull-request/2688)
+
+1. Logical Camera sees nested models
+    * [Pull request 2776](https://bitbucket.org/osrf/gazebo/pull-request/2776)
 
 ## Gazebo 7.8.1 (2017-06-08)